Minor changes to remedy typo-ish issues.

+++ Battle.java
@@ -205,10 +205,10 @@ public class Battle
 		nameRow += pl.getPlayerName() + " the " + pl.getPlayerClass() + spaces + en.getPlayerName() + " |";
 		System.out.println(nameRow);
 		
-		String plHP = String.valueOf(pl.getHP());
-		String plMaxHP = String.valueOf(pl.getMaxHealth());
-		String enHP = String.valueOf(en.getHP());
-		String enMaxHP = String.valueOf(en.getMaxHealth());
+		String plHP = String.valueOf((int)pl.getHP());
+		String plMaxHP = String.valueOf((int)pl.getMaxHealth());
+		String enHP = String.valueOf((int)en.getHP());
+		String enMaxHP = String.valueOf((int)en.getMaxHealth());
 		numSpace = 66 - plHP.length() - plMaxHP.length() - enHP.length() - enMaxHP.length();
 		spaces = "";
 		for(int i = 1; i <= numSpace; i++)
@@ -216,8 +216,8 @@ public class Battle
 		String hpRow = "| HP: " + plHP + "/" + plMaxHP + spaces + "HP: " + enHP + "/" + enMaxHP + " |";
 		System.out.println(hpRow);
 		
-		String plEP = String.valueOf(pl.getEP());
-		String plMaxEP = String.valueOf(pl.getMaxEP());
+		String plEP = String.valueOf((int)pl.getEP());
+		String plMaxEP = String.valueOf((int)pl.getMaxEP());
 		numSpace = 72 - plEP.length() - plMaxEP.length();
 		spaces = "";
 		for(int i = 1; i <= numSpace; i++)
@@ -232,7 +232,7 @@ public class Battle
 		String weapRow = "| Weapon: " + pl.getWeaponName() + spaces + "|";
 		System.out.println(weapRow);
 		
-		String plDam = String.valueOf(pl.getTotalRawDamage());
+		String plDam = String.valueOf((int)pl.getTotalRawDamage());
 		numSpace = 62 - plDam.length();
 		spaces = "";
 		for(int i = 1; i <= numSpace; i++)
